pyproject.toml
[build-system]
requires = [ "poetry-core>=2.0.1",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "django-oscar-cybersource"
version = "9.0.5"
description = "Integration between django-oscar and the Cybersource Secure Acceptance."
authors = ["thelab <thelabdev@thelab.co>"]
readme = "README.rst"
homepage = "https://gitlab.com/thelabnyc/django-oscar-cybersource"
repository = "https://gitlab.com/thelabnyc/django-oscar-cybersource"
license = "ISC"

[[tool.poetry.packages]]
include = "cybersource"
from = "src"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^4.2.18"
django-oscar = "3.2.5"
django-oscar-api = "3.3.0"
django-oscar-api-checkout = "^3.6.1"
thelabdb = "^0.2.0"
lxml = "^5.3.0"
phonenumbers = "^8.13.53"
python-dateutil = "^2.9.0"
django-stubs-ext = "^5.1.2"

[tool.poetry.group.dev.dependencies]
coverage = "^7.6.10"
flake8 = "^7.1.1"
instrumented-soap = "2.1.2"
psycopg2-binary = "^2.9.10"
PyYAML = "^6.0.2"
sorl-thumbnail = "^12.11.0"
sphinx-rtd-theme = "^3.0.2"
sphinx = "^8.1.3"
suds-community = "^1.2.0"
tox = "^4.24.1"
unittest-xml-reporting = "^3.2.0"
django-stubs = "^5.1.2"
mypy = "^1.14.1"
types-psycopg2 = "^2.9.21.20241019"
djangorestframework-stubs = "^3.15.2"
lxml-stubs = "^0.5.1"
types-python-dateutil = "^2.9.0.20241206"
types-pygments = "^2.19.0.20250107"

[tool.commitizen]
name = "cz_conventional_commits"
annotated_tag = true
gpg_sign = true
tag_format = "v$version"
update_changelog_on_bump = true
changelog_merge_prerelease = true
version_provider = "poetry"
version_scheme = "pep440"
version_files = [
    "pyproject.toml:version",
]
pre_bump_hooks = [
    "pre-commit run --all-files || true",
]
post_bump_hooks = [
    "git push origin master $CZ_POST_CURRENT_TAG_VERSION"
]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

# Strict mode, see mypy --help
warn_unused_configs = true
disallow_subclassing_any = true
disallow_any_generics = true
# disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
# warn_return_any = true
no_implicit_reexport = true
show_error_codes = true
# Not turned on by strict
strict_equality = true

[[tool.mypy.overrides]]
module = "oscar.*"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "oscarapi.*"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "suds.*"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "soap.*"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "*.tests.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "sandbox.settings"

[tool.isort]
profile = "black"
from_first = true

[tool.coverage.run]
branch = true
source_pkgs = ["oscarapicheckout"]

[tool.coverage.report]
show_missing = true
ignore_errors = true
